+++ app/Http/Controllers/ExpenseCategoryController.php
@@ -15,6 +15,10 @@ class ExpenseCategoryController extends Controller
 
 
     public function index(){
+        if(\Auth::user()->role->permissions->read!=1){
+            Toastr::warning(' You are not able to  access this Feature',"Access Denied");
+            return redirect()->back();
+        }
 
         $datas=ExpenseCategory::latest()->paginate(20);  
         return view('backend.expense_category.index',compact('datas'));
@@ -23,7 +27,9 @@ class ExpenseCategoryController extends Controller
 
     public function storeCategory(Request $request)
     {
-        
+        if(\Auth::user()->role->permissions->create!=1){
+            Toastr::warning(' You are not able to  access this Feature',"Access Denied");
+        }
         $request->validate([
 
             'category_name'=>'required|unique:expense_categories'    
@@ -42,7 +48,10 @@ class ExpenseCategoryController extends Controller
 
     public function UpdateCategory(Request $request,$id)
     {
-
+        if(\Auth::user()->role->permissions->update!=1){
+            Toastr::warning(' You are not able to  access this Feature',"Access Denied");
+            return redirect()->back();
+        }
         ExpenseCategory::find($id)->update($request->all());
         Toastr::info('Expense Category has been Upadted!',"Upadted");
         return redirect()->route('expense.category.list');
@@ -50,6 +59,11 @@ class ExpenseCategoryController extends Controller
     }
 
     public function DeleteCategory($id){
+
+        if(\Auth::user()->role->permissions->delete!=1){
+            Toastr::warning(' You are not able to  access this Feature',"Access Denied");
+            return redirect()->back();
+        }
 
         ExpenseCategory::find($id)->delete();
         Toastr::warning('Expense Category has been Deleted!',"Deleted");
